% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic_find.R
\docType{package}
\name{mosaic.find}
\alias{mosaic.find}
\alias{mosaic.find-package}
\title{mosaic.find}
\description{
Provides a function (mosaic_find()) designed to find rhythmic and non-rhythmic trends in multi-omics time course data using model selection and joint modeling, a method called MOSAIC (Multi-Omics Selection with Amplitude Independent Criteria).
}
\details{
To read more about this project and cite us, see MOSAIC: A Joint Modeling Methodology for Combined Circadian and Non-Circadian Analysis of Multi-Omics Data by H. De los Santos et al. (2020) doi: 10.1101/2020.04.27.064147
}
